properly perform check on existence of docker ID

+++ .ci/docker.sh
@@ -55,24 +55,30 @@ then
 
     DOCKER_CONTAINER_ID=$(docker ps | grep ${DOCKER_IMAGE} | awk '{print $1}')
     echo DOCKER_CONTAINER_ID=${DOCKER_CONTAINER_ID}
-    docker exec -ti ${DOCKER_CONTAINER_ID} /bin/bash -ec 'echo Testing build on docker for `cat /etc/system-release`'
-    docker logs $DOCKER_CONTAINER_ID
-    docker exec -ti ${DOCKER_CONTAINER_ID} /bin/bash -ec 'pip install -I --user "pip" "importlib" "codecov" "setuptools<38.2"'
-    docker exec -ti ${DOCKER_CONTAINER_ID} /bin/bash -ec 'python -c "import pkg_resources; print(pkg_resources.get_distribution('\''importlib'\''))"'
-    docker exec -ti ${DOCKER_CONTAINER_ID} /bin/bash -ec 'python -c "import pkg_resources; print(pkg_resources.get_distribution('\''pip'\''))"'
-    docker exec -ti ${DOCKER_CONTAINER_ID} /bin/bash -ec 'python -c "import pkg_resources; print(pkg_resources.get_distribution('\''setuptools'\''))"'
+    if [ -z {DOCKER_CONTAINER_ID+x} ];
+    then
+        docker exec -ti ${DOCKER_CONTAINER_ID} /bin/bash -ec 'echo Testing build on docker for `cat /etc/system-release`'
+        docker logs $DOCKER_CONTAINER_ID
+        docker exec -ti ${DOCKER_CONTAINER_ID} /bin/bash -ec 'pip install -I --user "pip" "importlib" "codecov" "setuptools<38.2"'
+        docker exec -ti ${DOCKER_CONTAINER_ID} /bin/bash -ec 'python -c "import pkg_resources; print(pkg_resources.get_distribution('\''importlib'\''))"'
+        docker exec -ti ${DOCKER_CONTAINER_ID} /bin/bash -ec 'python -c "import pkg_resources; print(pkg_resources.get_distribution('\''pip'\''))"'
+        docker exec -ti ${DOCKER_CONTAINER_ID} /bin/bash -ec 'python -c "import pkg_resources; print(pkg_resources.get_distribution('\''setuptools'\''))"'
+    fi
 else
     DOCKER_CONTAINER_ID=$(docker ps | grep ${DOCKER_IMAGE} | awk '{print $1}')
-    docker logs $DOCKER_CONTAINER_ID
+    if [ -z {DOCKER_CONTAINER_ID+x} ];
+    then
+        docker logs $DOCKER_CONTAINER_ID
 
-    if [ "${COMMAND}" = "stop" ]
-    then
-        docker exec -ti ${DOCKER_CONTAINER_ID} /bin/bash -ec "echo -ne \"------\nEND ${REPO_NAME} TESTS\n\";"
-        docker stop $DOCKER_CONTAINER_ID
-        docker rm -v $DOCKER_CONTAINER_ID
-    elif [ "${COMMAND}" = "other" ]
-    then
-        docker ps -a
+        if [ "${COMMAND}" = "stop" ]
+        then
+            docker exec -ti ${DOCKER_CONTAINER_ID} /bin/bash -ec "echo -ne \"------\nEND ${REPO_NAME} TESTS\n\";"
+            docker stop $DOCKER_CONTAINER_ID
+            docker rm -v $DOCKER_CONTAINER_ID
+        elif [ "${COMMAND}" = "other" ]
+        then
+            docker ps -a
+        fi
     fi
 fi
